backend/requirements.txt
aiohttp
openai
python-dotenv
torch
torchaudio
numpy
redis 
motor  
fastapi
uvicorn  

fastapi
uvicorn
slowapi
motor
redis
python-dotenv
torch
torchaudio
numpy
openai 
